Implement Age concern
  - Move migrated? and sponsored? into service objects so they can be
    called by the Age concern
  - Move age related methods to Age concern
  - Update user.rb to reference new service objects in a backwards
    compatible way
